Update .gitignore
Added  certain files to the ignore list.
Most of these files are assests that are not changing between branch and pull requests. It should be noted that the folders for the pygame2D that are ignored will now ignore any new files added in those folders, incase we add more enemies in the future. If we do want to add more of those assets we'll need to remove them from the .gitignore